Apply pr feedback
